+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 ## Change Log
 
+## v0.0.5
+- [feat]: Adds `signRelease` function to sign the release metadata.
+
+## v0.0.4
+- [chore]: Update `ov-types` to v0.0.6
+- [fix]: Fixed `userAgent.ts` and `resourceManager.ts` to use the new `IOVAgent` type.
+- [fix]: `walletManeger.ts` had a bug that prevented it from being used.
+
 ## v0.0.3 
 - [chore]: Update `ov-types` to v0.0.5
+
 ## v0.0.2
 - [feat]: Introduces 3 main agents, `UserAgent`, `PackageAgent`, and `ParentAgent` each with their own stores.
 - [feat]: Adds `ov-types` as a dependency.